Update and rename sorcery-speed.md to lotus-petal.md
The loop as previously written does not work for sorceries. Moreover, it misses the step to sacrifice lotus petal. A similar loop could be used for instants (not sorceries), but that loop is already covered in the "instant loop" section of gitrog.gg. I have changed this page to only focus on the lotus petal application and correct the errors mentioned.